requirements.txt
django
djangorestframework
pyyaml
requests
django-cors-headers
django-firebase-authentication
firebase_auth
pyrebase4
django-phonenumber-field
django-phonenumbers
Faker
